project/backend/requirements.txt
asgiref==3.8.1
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
environs==14.1.0
marshmallow==3.26.1
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2025.1
